define the budget details feature

+++ db/migrate/20200416005903_create_expenses.rb
@@ -4,7 +4,7 @@ class CreateExpenses < ActiveRecord::Migration[6.0]
       t.references :user, null: false, foreign_key: true
       t.integer :category
       t.date :date
-      t.float :amount
+      t.decimal :amount
 
       t.timestamps
     end
